+++ cloud-siem-20241212/ChangeLog.md
@@ -1,3 +1,9 @@
+2025-11-04 Version: 1.0.1
+- Update API CreateDetectionRule: add request parameters DetectionRuleTemplateId.
+- Update API CreateDetectionRule: add request parameters DetectionRuleTemplateVersion.
+- Update API ListIncidents: add response parameters Body.Incidents.$.IncidentTags.
+
+
 2025-10-14 Version: 1.0.0
 - Generated python 2024-12-12 for cloud-siem.